feat(storybook-addons): add Readme files and rename color-scheme addon
