+++ packages/storybook-addons/README.md
@@ -1,0 +1,11 @@
+# @vkontakte/storybook-addons
+
+Набор аддонов для Storybook, используемых в экосистеме VKUI.
+
+## Аддоны
+
+| Аддон | Описание | Документация                                   |
+|-------|----------|------------------------------------------------|
+| **Color Scheme** | Переключатель цветовой схемы (light/dark) в тулбаре | [README](src/addons/color-scheme/README.md)    |
+| **Storybook Theme** | Переключатель темы интерфейса Storybook | [README](src/addons/storybook-theme/README.md) |
+| **Source Button** | Кнопка ссылки на исходный код сториса | [README](src/addons/source-button/README.md)   |
