+++ packages/storybook-addons/src/addons/color-scheme/auto-register.ts
