+++ packages/storybook-addons/src/addons/color-scheme/constants.ts
@@ -1,0 +1,2 @@
+export const ADDON_ID = 'storybook/color-scheme';
+export const DEFAULT_PARAM_KEY = 'colorScheme';
