+++ packages/storybook-addons/src/addons/color-scheme/register.ts
