--- packages/storybook-addons/src/addons/colorScheme/constants.ts
@@ -1,2 +1,0 @@
-export const ADDON_ID = 'storybook/colorScheme';
-export const DEFAULT_PARAM_KEY = 'colorScheme';
